voiceai/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
python-dotenv==1.0.0
stripe==7.11.0
supabase==2.3.1
pydantic==2.5.3
torch==2.1.0
torchaudio==2.1.0